KristianRosenvold: Made RemoteProxy threadsafe.
This was quite serious, since the CleanUpThread was running totally unsynchronized, which just shows
the dangers of externalizing synchronization away from the structures that really should be
synchronized.

This fix is 'sinfully ugly' and to pay penance for my sins I fixed quite a lot of javadoc.

Technically the large number of subclasses to the RemoteProxy make safe & simple construction
almost impossible in a manner acceptable to any half-decent style police. I think changing
to a 'strategy' pattern that is constructor injected might be the way to fix this.

I still cannot guarantee the overall thread safety of the cleanup thread, so
this particulat thread is still not marked as 'reviewed'

git-svn-id: https://selenium.googlecode.com/svn/trunk@13638 07704840-8298-11de-bf8c-fd130f914ac9