Fixed an issue whereby a Selenium RC TestSession without an external key could never be cleaned up.
These tests are pretty weak, but illustrate that an exception is not thrown.  I need to figure out how to best mock this to verify behavior.  I suspect this means extracting an interface because I can't figure out how to do a partial mock with the version of EasyMock (2.4) we have locked down.

git-svn-id: https://selenium.googlecode.com/svn/trunk@13771 07704840-8298-11de-bf8c-fd130f914ac9