KristianRosenvold: Added a 3 hour SO_TIMEOUT to all httpclients
3 hours is the implementation value of "ages", where the current timeout
is "infinity". After "ages" we fail, no matter what.

This timeout governs *all* http requests originating from java code, between client,
grid, server and browsers. In reality, 60 seconds is the most sensible value,
except when the client is connecting to a grid, since the grid uses long-running
requests as a means to keep a queue of clients in wait. So if anyone is actually
willing to wait /more/ than three hours to get a grid node, this has to change.

This timeout means our CI environment no longer will hang for three *days*
without timing out a build when there are some nodes that have simply
taken a long walk on a short pier.

Personally I think an hour would be even better, or even less. But then
it would probably have to be configurable.

All over, I wish I had a somewhat better way of dealing with this.
The only way I can think of is to somehow two-phase the grid queuing
process, so the client could know up-front it is entering a queue.

git-svn-id: https://selenium.googlecode.com/svn/trunk@14737 07704840-8298-11de-bf8c-fd130f914ac9